% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_sig.R
\name{ks_sig}
\alias{ks_sig}
\title{ks_sig}
\usage{
ks_sig(ks.pval.unadj)
}
\arguments{
\item{ks.pval.unadj}{The output from \code{ks_test} which is a list
of p-values from the KS test with gene names.}

\item{cexpr}{A dataframe that contains the covariate values.
The rows of the dataframe are the corresponding samples/cells from the counts
matrix from \code{filter_counts}.
The cells of the dataframe are the covariates to be included in the GLM.}
}
\value{
List object containing the significant gene indices from the KS test,
their adjusted p-values
}
\description{
This function is used to select genes significant from the
\code{ks_test}.
}
