% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo}
\alias{logo}
\title{A Statistical Framework for Modeling and Identifying Differential Distributions in Single-cell RNA-sequencing Data}
\usage{
logo()
}
\value{
... Reports the logo of the package:scShapes.
}
\description{
A Statistical Framework for Modeling and Identifying Differential Distributions in Single-cell RNA-sequencing Data
}
