% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_param.R
\name{model_param}
\alias{model_param}
\title{model_param}
\usage{
model_param(fit.model, gof.res, model = NULL)
}
\arguments{
\item{fit.model}{A list of models fitted by 'glm' from \code{fit_models}.}

\item{gof.res}{A list of selected model distributions for genes \code{select_model}.}

\item{model}{A specific model to fit (1:P, 2:NB, 3:ZIP, 4:ZINB, NULL:All)}
}
\value{
A list of models fitted by 'glm'
}
\description{
This function returns model parameters based on the best fit distribution
as selected by \code{distr_fit} and models fitted by \code{fit_models}
}
