% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{A Statistical Framework for Modeling and Identifying Differential Distributions in Single-cell RNA-sequencing Data}
\usage{
version()
}
\value{
... Reports the version of the package:scShapes.
}
\description{
A Statistical Framework for Modeling and Identifying Differential Distributions in Single-cell RNA-sequencing Data
}
