% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bic.R
\name{model_bic}
\alias{model_bic}
\title{model_bic}
\usage{
model_bic(fit_models)
}
\arguments{
\item{fit_list}{A list of models fitted from \code{fit_models}}
}
\value{
A dataframe containing the BIC values for each distribution type (P, NB, ZIP, ZINB).
}
\description{
This function is used to calculate the Bayesian information criterion of the models fitted in
\code{fit_models}.
}
